requirements.txt
# Core
streamlit>=1.35.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=2.2.0

# Browser-based OKTA/SSO login
selenium>=4.21.0
webdriver-manager>=4.0.0

# Charts
plotly>=5.22.0
